+++ CHANGES.md
@@ -1,6 +1,7 @@
 # About CrateDB changelog
 
 ## Unreleased
+- Prompt: Added `instructions-general.md` file when generating bundle
 
 ## v0.0.6 - 2025-07-21
 - Prompt: Added instructions about working with CrateDB to be used for